Fix overflow in Memory.Experimental.MacAvailableMemoryPercentFreePageCache2 calculation
info.file_backed is an int and 1024 * INT_MAX may overflow. Indeed this
UBSan failure shows a value of 16146996, or about 15 GiB. That does not
fit in int, which can only represent up to 2 GiB. As all the rest of the
values are uint64_t, cast to uint64_t before performing the computation.

See the following UBSan error:
https://ci.chromium.org/ui/labs/p/chromium/inv/build-8744198296393197089/test/ninja%3A%2F%2Fchrome%2Ftest%3Aunit_tests%2FPerformanceManagerMetricsProviderDesktopTest.TestBothModes/variant/989b21184728cff2?resultIndex=0

Bug: 40248746
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5650019
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1319167}

+++ chrome/browser/performance_manager/metrics/metrics_provider_desktop.cc
@@ -6,6 +6,7 @@
 
 #include "base/metrics/histogram_functions.h"
 #include "base/metrics/histogram_macros.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/power_monitor/cpu_frequency_utils.h"
 #include "base/process/process_metrics.h"
 #include "base/system/sys_info.h"
@@ -443,7 +444,9 @@ void MetricsProviderDesktop::RecordAvailableMemoryMetrics() {
     // bytes
     base::UmaHistogramPercentage(
         "Memory.Experimental.MacAvailableMemoryPercentFreePageCache2",
-        (available_bytes + (info.file_backed * 1024)) * 100 / total_bytes);
+        (available_bytes +
+         (base::checked_cast<uint64_t>(info.file_backed) * 1024u)) *
+            100u / total_bytes);
   }
 #endif
 }
